#5 Move all portfolio_factory.py related classes to portfolio_factory directory, expose base.py as entry point
