+++ backend/src/models/portfolio_factory/base.py
@@ -1,0 +1,24 @@
+from backend.src.api import *
+from backend.src.models import portfolio
+from .trading212 import Trading212PortfolioFactory
+
+
+class PortfolioFactory:
+  """Factory pattern to return a Portfolio instance depending on the Broker API as input"""
+
+  @staticmethod
+  def build(broker_api):
+    """
+    Build a Portfolio instance using the Broker API
+
+    Args:
+      broker_api (object): The Broker API instance
+
+    Returns:
+      object: The Portfolio instance
+    """
+
+    if isinstance(broker_api, Trading212):
+      return Trading212PortfolioFactory.build(broker_api)
+    else:
+      raise ValueError(f"Unknown broker api parameter: {broker_api}")
